Изменил тип с float на int

+++ src/sprites/tile.py
@@ -100,7 +100,7 @@ class Tile(Sprite):
         ])
 
     @staticmethod
-    def get_half_of_size(tile_size: int, pixel_size: int, perspective_angle: int, camera_distance: float) -> tuple[
+    def get_half_of_size(tile_size: int, pixel_size: int, perspective_angle: int, camera_distance: int) -> tuple[
         int, int]:
         return (round(tile_size * pixel_size * camera_distance / 10 * cos(radians(perspective_angle))),
                 round(tile_size * pixel_size * camera_distance / 10 * sin(radians(perspective_angle))))
